requirements.txt
appdirs==1.4.3
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.11
Django==1.11
django-appconf==1.0.2
django-bootstrap-toolkit==2.15.0
django-compressor==2.1.1
django-extensions==1.7.8
django-localflavor
django-phonenumber-field
django-selectable==1.0.0
enum34==1.1.6
ipython==5.3.0
ipython-genutils==0.2.0
packaging==16.8
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-phonenumbers
pytz==2017.2
rcssmin==1.0.6
rjsmin==1.0.12
scandir==1.5
simplegeneric==0.8.1
singledispatch==3.4.0.3
sqlparse
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
wrapt
setproctitle
